refactor(ingest): one module per derived artifact — app.ingest.need_map
``consolidate`` was a verb borrowed from the eval pipeline's S3, and it named
the act rather than the thing. The module produces the need map, so it is named
for that, matching the table it writes and the artifact it returns.

Folded ``topics.py`` in rather than renaming beside it. It had exactly one
consumer — this module — and nothing sits between the two halves: clustering
feeds naming and nobody else calls it. It was also misnamed, producing clusters
rather than topics, which only exist after the step it did not do. One module
per derived artifact is how entity_types is organised (extract, fold, name,
merge, 863 lines); this is 476.

The halves stay separate functions and separately tested, so the property
topics.py's docstring argued for — that the grouping is deterministic and
inspectable on its own — is unchanged. It was about coupling, not files.

  cluster_needs   deterministic, no LLM
  name_cluster    one LLM call per cluster
  run_derivation  load -> cluster -> name -> record

``run_consolidation`` becomes ``run_derivation``, matching entity_types' verb
for the same kind of whole-corpus pass, and the fan-out helper becomes
``name_clusters`` so it is not one letter from the module.

The db store is imported as ``store`` inside the ingest module of the same
name — legal either way, since Python does not bind a module's own name in its
namespace, but ``store.record(...)`` cannot be misread as a self-call.
